+++ kotlin/tests/src/test/kotlin/com/squareup/moshi/kotlin/reflect/-MoshiKotlinExtensions.kt
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2019 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.squareup.moshi.kotlin.reflect
 
 import com.squareup.moshi.JsonAdapter
@@ -51,12 +66,16 @@ internal fun KType.typeAnnotations(): Set<Annotation> {
 }
 
 @PublishedApi
-internal fun KType.toType(): Type {
+internal fun KType.toType(allowPrimitives: Boolean = true): Type {
   classifier?.let {
     when (it) {
       is KTypeParameter -> throw IllegalArgumentException("Type parameters are not supported")
       is KClass<*> -> {
-        val javaType = it.java
+        val javaType = if (allowPrimitives) {
+          it.java
+        } else {
+          it.javaObjectType
+        }
         if (javaType.isArray) {
           return Types.arrayOf(javaType.componentType)
         }
@@ -82,7 +101,7 @@ internal fun KType.toType(): Type {
 }
 
 internal fun KTypeProjection.toType(): Type {
-  val javaType = type?.toType() ?: return Any::class.java
+  val javaType = type?.toType(allowPrimitives = false) ?: return Any::class.java
   return when (variance) {
     null -> Any::class.java
     KVariance.INVARIANT -> javaType